pyproject.toml
[tool.poetry]
name = "ngic-whisper"
version = "0.1.0"
description = "NGIC Whisper ASR Test Api"
authors = ["Christopher Biggs <christopher.l.biggs@gmail.com>"]
readme = "README.md"
packages = [
    {include = "app"}
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.94.0"
unidecode = "^1.3.6"
uvicorn = {version = "^0.21.0", extras = ["standard"]}
gunicorn = "^20.1.0"
tqdm = "^4.65.0"
transformers = "^4.26.1"
python-multipart = "^0.0.6"
ffmpeg-python = "^0.2.0"
torch = "^1.13.1"
faster-whisper = {git = "https://github.com/guillaumekln/faster-whisper.git"}
fastapi-offline-swagger-ui = {git = "https://github.com/ahmetoner/fastapi-offline-swagger-ui"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
